make the metrictank entrypoint script generic
This renames scripts/metrictank_entrypoint.sh to
scripts/wait_for_endpoint.sh and makes it more generic so it can be
reused in other use cases too.

It also improves the way how the script checks for open ports. One
problem of the previous version was that it only tested if a given port
is open, as in something is listening, but it did not check if the
connection can be established and stays up.
This matters because docker-proxy might start listening on a tcp port
before the service that backs the proxy is ready. In these cases it will
accept incoming connections but close them right away.  With the changes
in this commit the port is only considered "up" if the connection stays
up for 3 seconds.