Adds an end2end test script
This script uses docker-compose to start a full functioning environment.
The goal is to test the functionality of the docker configuration and
all the involved pieces, if this test passes it can be assumed that the
services are up and functioning.
It uses the Fakemetrics script to generate statistics and send them
to Metrictank's carbon input. After doing that for a while it queries
the graphite API how many metrics have been received by Metrictank, if
the number of metrics per time is at least what Fakemetrics has
generated, then the test passes.

Note:
end2end_test.sh downloads version 1.8.1 of docker-compose. This is
necessary because the version on the Circle CI images is not recent
enough to use our version 2 configuration file.
Once Circle CI updates the images we can get rid of that step.